requirements.txt
aiofiles==23.1.0
anyio==3.7.1
click==8.1.6
exceptiongroup==1.1.2
graphql-core==3.2.3
h11==0.14.0
html5tagger==1.3.0
httptools==0.6.0
idna==3.4
libcst==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
peewee==3.16.2
Pygments==2.15.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.4.2
sanic==23.6.0
sanic-routing==23.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.31.0
strawberry-graphql==0.196.2
tracerite==1.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.1
uvloop==0.17.0
websockets==11.0.3
